fix(vmconnect): make client support opt-in
Gate VMConnect configuration and handshake code behind an explicit client feature, expose the supported facade boundary, and move public-contract coverage into the integration suite.

Document the empty-username cookie behavior and make the public console mode extensible.
