+++ crates/ironrdp-client/Cargo.toml
@@ -39,6 +39,7 @@ qoi = ["ironrdp-connector/qoi", "ironrdp-session/qoi"]
 qoiz = ["ironrdp-connector/qoiz", "ironrdp-session/qoiz"]
 dvc-pipe-proxy = ["dep:ironrdp-dvc-pipe-proxy"]
 dvc-com-plugin = ["dep:ironrdp-dvc-com-plugin"]
+vmconnect = ["dep:ironrdp-vmconnect"]
 
 all = [
     "sound",
@@ -48,6 +49,7 @@ all = [
     "gateway",
     "dvc-pipe-proxy",
     "dvc-com-plugin",
+    "vmconnect",
 ]
 
 [dependencies]
@@ -64,7 +66,7 @@ ironrdp-echo = { path = "../ironrdp-echo", version = "0.4" }
 ironrdp-tls = { path = "../ironrdp-tls", version = "0.2" } # public
 ironrdp-tokio = { path = "../ironrdp-tokio", version = "0.10", features = ["reqwest"] }
 ironrdp-rdcleanpath = { path = "../ironrdp-rdcleanpath", version = "0.2" }
-ironrdp-vmconnect = { path = "../ironrdp-vmconnect", version = "0.1" }
+ironrdp-vmconnect = { path = "../ironrdp-vmconnect", version = "0.1", optional = true }
 ironrdp-cfg = { path = "../ironrdp-cfg", version = "0.1" }
 ironrdp-propertyset = { path = "../ironrdp-propertyset", version = "0.1" } # public
