+++ crates/ironrdp-client/src/rdp.rs
@@ -748,6 +748,7 @@ async fn connect_direct(
     let stream = TcpStream::connect(&dest)
         .await
         .map_err(|e| ironrdp_connector::custom_err!("TCP connect", e))?;
+    #[cfg(feature = "vmconnect")]
     let pcb_deadline = tokio::time::Instant::now() + ironrdp_vmconnect::PCB_TRANSMIT_DEADLINE;
     let client_addr = stream
         .local_addr()
@@ -755,11 +756,11 @@ async fn connect_direct(
     let framed = ironrdp_tokio::TokioFramed::new(stream);
 
     let connector = build_connector(config, client_addr, input_sender, cliprdr_factory);
+    #[cfg(feature = "vmconnect")]
     if config.vm_id().is_some() {
-        vmconnect_handshake_and_finalize(framed, connector, config, pcb_deadline).await
-    } else {
-        tls_handshake_and_finalize(framed, connector, config).await
+        return vmconnect_handshake_and_finalize(framed, connector, config, pcb_deadline).await;
     }
+    tls_handshake_and_finalize(framed, connector, config).await
 }
 
 /// RDS gateway TCP → gateway auth → TLS connection.
@@ -773,6 +774,7 @@ async fn connect_gateway(
     use ironrdp_mstsgu::GwConnectTarget;
 
     // VMConnect needs destination port 2179; GwConnectTarget does not carry it yet (TODO below).
+    #[cfg(feature = "vmconnect")]
     if config.vm_id().is_some() {
         return Err(ironrdp_connector::general_err!(
             "vmconnect cannot be used over an RDS gateway until the target port is propagated"
@@ -908,6 +910,7 @@ where
 }
 
 /// Hyper-V console connect via ironrdp-vmconnect, then shared RDP tail.
+#[cfg(feature = "vmconnect")]
 async fn vmconnect_handshake_and_finalize<S>(
     mut framed: ironrdp_tokio::TokioFramed<S>,
     mut connector: ironrdp_connector::ClientConnector,
