+++ crates/ironrdp-testsuite-extra/tests/main.rs
@@ -5,3 +5,4 @@ mod agent;
 mod client_config;
 mod dvc_pipe_proxy;
 mod e2e;
+mod vmconnect;
