+++ crates/ironrdp-vmconnect/Cargo.toml
@@ -14,6 +14,7 @@ categories.workspace = true
 
 [lib]
 doctest = false
+test = false
 
 [dependencies]
 ironrdp-async = { path = "../ironrdp-async", version = "0.10" } # public
