+++ crates/ironrdp-vmconnect/src/lib.rs
@@ -43,6 +43,7 @@ const POST_CREDSSP_PROTOCOL: SecurityProtocol = SecurityProtocol::HYBRID;
 const ENHANCED_MODE_SUFFIX: &str = ";EnhancedMode=1";
 
 /// Hyper-V VM console connection mode.
+#[non_exhaustive]
 #[derive(Clone, Copy, Debug, Default, PartialEq, Eq)]
 pub enum Mode {
     /// Use the guest's Enhanced Session RDP backend.
@@ -192,154 +193,5 @@ fn ensure_selected_credssp(state: &ClientConnectorState) -> ConnectorResult<()> 
             "Initiation",
             "server must select HYBRID for a Hyper-V console, but it selected {selected}",
         ))
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use ironrdp_connector::{ClientConnector, ClientConnectorState, Credentials};
-    use ironrdp_core::{WriteBuf, decode};
-    use ironrdp_pdu::nego::{ConnectionRequest, SecurityProtocol};
-    use ironrdp_pdu::pcb::PreconnectionBlob;
-    use ironrdp_pdu::x224::X224;
-
-    use super::{Mode, encode_preconnection_blob, ensure_selected_credssp, prepare_connector};
-
-    #[test]
-    fn pcb_payload_selects_console_mode() {
-        const VM_ID: &str = "efd1efab-c750-4262-b1bb-af0f7733bdd6";
-
-        for (mode, expected) in [
-            (Mode::Enhanced, format!("{VM_ID};EnhancedMode=1")),
-            (Mode::Basic, VM_ID.to_owned()),
-        ] {
-            let bytes = encode_preconnection_blob(VM_ID, mode).expect("encode");
-            let pcb: PreconnectionBlob = decode(&bytes).expect("decode");
-
-            assert_eq!(pcb.v2_payload.as_deref(), Some(expected.as_str()));
-        }
-    }
-
-    #[test]
-    fn accepts_hybrid_selected_after_enhanced_mode_authentication() {
-        let state = ClientConnectorState::EnhancedSecurityUpgrade {
-            selected_protocol: SecurityProtocol::HYBRID,
-        };
-
-        ensure_selected_credssp(&state).expect("CredSSP protocol");
-    }
-
-    #[test]
-    fn rejects_hybrid_ex_selected_after_enhanced_mode_authentication() {
-        let state = ClientConnectorState::EnhancedSecurityUpgrade {
-            selected_protocol: SecurityProtocol::HYBRID_EX,
-        };
-
-        ensure_selected_credssp(&state).expect_err("HYBRID_EX must be rejected");
-    }
-
-    #[test]
-    fn pcb_transmit_deadline_matches_ms_rdpeps() {
-        assert_eq!(super::PCB_TRANSMIT_DEADLINE, core::time::Duration::from_secs(10));
-    }
-
-    fn test_connector(enable_tls: bool, enable_credssp: bool) -> ClientConnector {
-        use ironrdp_pdu::gcc;
-        use ironrdp_pdu::rdp::capability_sets::MajorPlatformType;
-
-        ClientConnector::new(
-            ironrdp_connector::Config {
-                desktop_size: ironrdp_connector::DesktopSize {
-                    width: 800,
-                    height: 600,
-                },
-                desktop_scale_factor: 0,
-                enable_tls,
-                enable_credssp,
-                credentials: Credentials::UsernamePassword {
-                    username: "u".into(),
-                    password: "p".into(),
-                },
-                domain: None,
-                client_build: 0,
-                client_name: "test".into(),
-                keyboard_type: gcc::KeyboardType::IbmEnhanced,
-                keyboard_subtype: 0,
-                keyboard_layout: 0,
-                keyboard_functional_keys_count: 12,
-                connection_type: gcc::ConnectionType::Lan,
-                ime_file_name: String::new(),
-                bitmap: None,
-                dig_product_id: String::new(),
-                client_dir: String::new(),
-                platform: MajorPlatformType::UNSPECIFIED,
-                hardware_id: None,
-                request_data: None,
-                autologon: false,
-                enable_audio_playback: false,
-                license_cache: None,
-                compression_type: None,
-                enable_server_pointer: false,
-                pointer_software_rendering: false,
-                multitransport_flags: None,
-                performance_flags: Default::default(),
-                timezone_info: Default::default(),
-                alternate_shell: String::new(),
-                work_dir: String::new(),
-            },
-            "127.0.0.1:2179".parse().unwrap(),
-        )
-    }
-
-    #[test]
-    fn prepare_connector_accepts_tls_and_credssp() {
-        prepare_connector(&mut test_connector(true, true)).expect("required flags set");
-    }
-
-    #[test]
-    fn prepare_connector_rejects_disabled_tls() {
-        let err = prepare_connector(&mut test_connector(false, true)).expect_err("TLS required");
-        assert!(err.to_string().contains("TLS"), "{err}");
-    }
-
-    #[test]
-    fn prepare_connector_rejects_disabled_credssp() {
-        let err = prepare_connector(&mut test_connector(true, false)).expect_err("CredSSP required");
-        assert!(err.to_string().contains("CredSSP"), "{err}");
-    }
-
-    #[test]
-    fn prepare_connector_advertises_hybrid_without_hybrid_ex() {
-        let mut connector = test_connector(true, true);
-        prepare_connector(&mut connector).expect("prepare connector");
-
-        let mut output = WriteBuf::new();
-        connector
-            .initiate_with_security_protocol(SecurityProtocol::HYBRID, &mut output)
-            .expect("encode X.224 request");
-        let X224(request): X224<ConnectionRequest> = decode(output.filled()).expect("decode X.224 request");
-
-        assert_eq!(request.protocol, SecurityProtocol::HYBRID);
-    }
-
-    #[test]
-    fn guest_rdp_sequence_omits_host_username_cookie() {
-        let mut connector = test_connector(true, true);
-        prepare_connector(&mut connector).expect("prepare connector");
-        connector.config.credentials = Credentials::UsernamePassword {
-            username: String::new(),
-            password: String::new(),
-        };
-        connector.config.domain = None;
-        connector.config.autologon = false;
-
-        let mut output = WriteBuf::new();
-        connector
-            .initiate_with_security_protocol(SecurityProtocol::HYBRID, &mut output)
-            .expect("encode X.224 request");
-        let X224(request): X224<ConnectionRequest> = decode(output.filled()).expect("decode X.224 request");
-
-        assert_eq!(request.protocol, SecurityProtocol::HYBRID);
-        assert_eq!(request.nego_data, None);
     }
 }
